requirements.txt
bandit==1.7.5
filesplit==4.0.1
flake8==6.0.0
jsonschema==4.17.3
numpy==1.21.5
pandas==2.0.2
paramiko==2.9.0
paramiko-expect==0.3.5
platformdirs==3.2.0
pyshark==0.6
pytest==7.3.1
pytest-html==3.2.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.32.3
ruamel.yaml==0.18.5
scp==0.14.5
setuptools==70.0.0
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.19.1